Tweaked Example and Hillclimbing Weighting
Tweaked example so it actually makes progress. Also, added an option for
the simulation to stop at MECO, as allowing simulation to continue
beyond MECO unfairly disadvantages orbits with slightly-too-low
perigees. This is the last update to the MATLAB version, as MATLAB is
terrible and so I'm moving to SciPy,